ref(settings): Move settings files to a module (#1829)
