+++ snuba/settings/settings_test.py
