+++ snuba/settings/validation.py
@@ -1,0 +1,6 @@
+from typing import Any, Mapping
+
+
+def _validate_settings(locals: Mapping[str, Any]) -> None:
+    # TODO: Add validation
+    pass
