--- snuba/settings_base.py
@@ -1,5 +1,0 @@
-# Do not remove this file. Previous versions of the settings file format
-# needed to import this, and custom settings files may still expect it.
-import warnings
-
-warnings.warn("Settings files should not import settings_base.py", DeprecationWarning)
